MOJ-95 Use native WebSocket as provided by the browser
During the registration of the end-points in WebSocketConfiguration,
withSockJS() is still called. withSockJS() causes the 'raw' end-point to
be post-fixed with /websocket. Without withSockJS, the end-point seems
to be missing.

Unfortunately, WebSockets require a FQDN, which is why the definition of
the end-point in javascript is now a bit less elegant.